+++ components/chips/color-chips/ColorChips.tsx
@@ -1,0 +1,12 @@
+import ColorChip from './ColorChip';
+import ColorChipsStyle from './ColorChips.style';
+
+const ColorChips = () => {
+  return (
+    <ColorChipsStyle>
+      <ColorChip />
+    </ColorChipsStyle>
+  );
+};
+
+export default ColorChips;
